requirements.txt
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
matplotlib>=3.8.4
seaborn>=0.13.2
tensorflow-cpu>=2.12,<3.0
scipy>=1.13.0
streamlit>=1.36.0
plotly>=5.22.0

